requirements.txt
tiktoken
openai[embeddings, wandb, datalib]
PyPDF2
logging
scikit-learn
python-dotenv
chromadb[config, utils]